requirements.txt
tda-api
webdriver-manager
pytest
